Add thread churning regression test for issue#337
This adds a test using @acj's example thread_churn.rb to test that the retry
logic around process locking allows for the process to still be successfully
locked.

It runs 1000 samples, every 1ms, for approximating recording with 1000hz sampling,
to make for a particularly pathological test.

The success criteria are based on those from `fn record` in main.rs, mimicking
the maximum error and maximum error rate we should expect, as it is still
possible to have other MemoryCopyErrors.

If any MemoryCopyError::ProcessEnded is detected, however, the